changes related to qtree
