+++ examples/qtree_get_example.go
@@ -7,32 +7,26 @@ import (
 )
 
 func main() {
-	// Initialize the ONTAP client
 	client := ontap.NewClient(
-		"http://192.168.200.60",
+		"https://mycluster.example.com",
 		&ontap.ClientOptions{
-			Version:           "1.32", // Ensure this version is compatible with your ONTAP system
-			BasicAuthUser:     "umonitor",
-			BasicAuthPassword: "sxZeJs4n",
+			Version:           "1.160",
+			BasicAuthUser:     "admin",
+			BasicAuthPassword: "secret",
 			SSLVerify:         false,
 			Debug:             false, // Enable debugging
 			Timeout:           60 * time.Second,
 		},
 	)
-
-	// Fetch Qtree information iteratively
-	qtreeOptions := &ontap.QtreeListIterOptions{
+	qtreeOptions := &ontap.QtreeGetIterOptions{
 		MaxRecords: 1024,
 	}
-
-	var allQtrees []*ontap.QtreeListIterResponse
-	qtreeResponses, err := client.QtreeListIterAPI(qtreeOptions)
+	qtreeResponses, err := client.QtreeGetIterAPI(qtreeOptions)
 	if err != nil {
 		fmt.Printf("Error fetching Qtree information: %v\n", err)
 		return
 	}
 
-	// Process and display the Qtree information
 	if len(qtreeResponses) > 0 {
 		fmt.Println("Qtree Information:")
 		for _, response := range qtreeResponses {
@@ -45,11 +39,8 @@ func main() {
 				fmt.Printf("Security Style: %s\n", qtree.SecurityStyle)
 				fmt.Println()
 			}
-			allQtrees = append(allQtrees, response)
 			fmt.Printf("Total Qtrees: %d\n", response.Results.NumRecords)
-
 		}
-		//fmt.Printf("Total Qtrees: %d\n", len(allQtrees))
 	} else {
 		fmt.Println("No Qtrees found.")
 	}
